services/gateway/requirements.txt
Django>=4.2,<5
django-ninja>=1.0
django-cors-headers>=4.3
PyJWT>=2.8
channels>=4.0
channels-redis>=4.2
daphne>=4.0
requests>=2.31
boto3>=1.34
pika>=1.3
gunicorn>=21.0
